requirements.txt
Flask>=2.2
Authlib>=1.2.0
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0
python-dotenv>=0.21.0
qrcode>=7.3
pillow>=10.0
